Function odbc::ffi::SQLDriverConnectW

pub unsafe extern "system" fn SQLDriverConnectW(
    connection_handle: *mut Dbc, 
    window_handle: *mut c_void, 
    in_connection_string: *const u16, 
    string_length_1: i16, 
    out_connection_string: *mut u16, 
    buffer_length: i16, 
    string_length_2: *mut i16, 
    driver_completion: SqlDriverConnectOption
) -> SQLRETURN

An alternative to SQLConnect. It supports data sources that require more connection information than the three arguments in SQLConnect, dialog boxes to prompt the user for all connection information, and data sources that are not defined in the system information

Returns

SQL_SUCCESS, SQL_SUCCESS_WITH_INFO, SQL_ERROR, SQL_INVALID_HANDLE, SQL_NO_DATA, or SQL_STILL_EXECUTING