Function odbc::ffi::SQLFetch

pub unsafe extern "system" fn SQLFetch(statement_handle: *mut Stmt) -> SQLRETURN

SQLFetch fetches the next rowset of data from the result set and returns data for all bound columns.

Returns

SQL_SUCCESS, SQL_SUCCESS_WITH_INFO, SQL_ERROR, SQL_INVALID_HANDLE, SQL_NO_DATA or SQL_STILL_EXECUTING