Function odbc::ffi::SQLForeignKeysW

pub unsafe extern "system" fn SQLForeignKeysW(
    statement_handle: *mut Stmt, 
    pk_catalog_name: *const u16, 
    pk_catalog_name_length: i16, 
    pk_schema_name: *const u16, 
    pk_schema_name_length: i16, 
    pk_table_name: *const u16, 
    pk_table_name_length: i16, 
    fk_catalog_name: *const u16, 
    fk_catalog_name_length: i16, 
    fk_schema_name: *const u16, 
    fk_schema_name_length: i16, 
    fk_table_name: *const u16, 
    fk_table_name_length: i16
) -> SQLRETURN

Can return:

A list of foreign keys in the specified table (columns in the specified table that refer to primary keys in other tables).

A list of foreign keys in other tables that refer to the primary key in the specified table.

The driver returns each list as a result set on the specified statement.

Returns

SQL_SUCCESS, SQL_SUCCESS_WITH_INFO, SQL_ERROR, SQL_INVALID_HANDLE, or SQL_STILL_EXECUTING.