Struct odbc::DiagRec

pub struct DiagRec {
    pub state: [SQLCHAR; 6],
    pub native_error_pointer: SQLINTEGER,
    pub message: String,
}

ODBC Diagnostic Record

Fields
state: [SQLCHAR; 6] native_error_pointer: SQLINTEGER message: String
Methods

impl DiagRec

unsafe fn create(handle_type: SQLSMALLINT, handle: SQLHANDLE) -> DiagRec

Reads first diagnostic record from handle

This function will panic if no Diagnostic record is available. Its primary use is to create one after another ODBC function returned SQL_ERROR