Struct odbc::Environment

pub struct Environment<V> { /* fields omitted */ }

Handle to an ODBC Environment

Creating an instance of this type is the first thing you do then using ODBC. The environment must outlive all connections created with it

Methods

impl Environment<Version3>

fn drivers(&self) -> Result<Vec<DriverInfo>>

Stores all driver description and attributes in a Vec

fn data_sources(&self) -> Result<Vec<DataSourceInfo>>

Stores all data source server names and descriptions in a Vec

fn system_data_sources(&self) -> Result<Vec<DataSourceInfo>>

Stores all sytem data source server names and descriptions in a Vec

fn user_data_sources(&self) -> Result<Vec<DataSourceInfo>>

Stores all user data source server names and descriptions in a Vec

impl Environment<NoVersion>

fn new() -> Result<Environment<NoVersion>, EnvAllocError>

Allocates a new ODBC Environment

Declares the Application's ODBC Version to be 3

fn set_odbc_version_3(self) -> Result<Environment<Version3>>

Tells the driver(s) that we will use features of up to ODBC version 3