Struct odbc::Statement

pub struct Statement<'a, S> { /* fields omitted */ }

RAII wrapper around ODBC statement

Methods

impl<'a> Statement<'a, Allocated>

fn with_parent(ds: &'a DataSource<Connected>) -> Result<Self>

fn tables<'b>(self) -> Result<Statement<'a, Executed>>

fn exec_direct(self, statement_text: &str) -> Result<Statement<'a, Executed>>

Executes a preparable statement, using the current values of the parameter marker variables if any parameters exist in the statement.

SQLExecDirect is the fastest way to submit an SQL statement for one-time execution.

impl<'a> Statement<'a, Executed>

fn num_result_cols(&self) -> Result<i16>

The number of columns in a result set

Can be called successfully only when the statement is in the prepared, executed, or positioned state. If the statement does not return columns the result will be 0.

fn fetch<'b>(&'b mut self) -> Result<Option<Cursor<'b, 'a>>>

Fetches the next rowset of data from the result set and returns data for all bound columns.

Return

Returns false on the last row