Struct odbc::Connection

pub struct Connection<'env, AC: AutocommitMode> { /* fields omitted */ }

Represents a connection to an ODBC data source

Methods

impl<'env> Connection<'env, AutocommitOn>

pub fn disable_autocommit(self) -> Result<Connection<'env, AutocommitOff>, Self>

impl<'env> Connection<'env, AutocommitOff>

pub fn enable_autocommit(self) -> Result<Connection<'env, AutocommitOn>, Self>

pub fn commit(&mut self) -> Result<()>

pub fn rollback(&mut self) -> Result<()>

impl<'env, AC: AutocommitMode> Connection<'env, AC>

pub fn is_read_only(&mut self) -> Result<bool>

true if the data source is set to READ ONLY mode, false otherwise.

This characteristic pertains only to the data source itself; it is not characteristic of the driver that enables access to the data source. A driver that is read/write can be used with a data source that is read-only. If a driver is read-only, all of its data sources must be read-only.

pub fn disconnect(self) -> Result<()>

Closes the connection to the data source. If not called explicitly the disconnect will be invoked implicitly by drop()