Struct odbc::Environment

pub struct Environment<V> { /* fields omitted */ }

Handle to an ODBC Environment

Creating an instance of this type is the first thing you do then using ODBC. The environment must outlive all connections created with it.

Methods

impl Environment<Version3>

pub fn drivers(&mut self) -> Result<Vec<DriverInfo>>

Stores all driver description and attributes in a Vec

pub fn data_sources(&mut self) -> Result<Vec<DataSourceInfo>>

Stores all data source server names and descriptions in a Vec

pub fn system_data_sources(&mut self) -> Result<Vec<DataSourceInfo>>

Stores all system data source server names and descriptions in a Vec

pub fn user_data_sources(&mut self) -> Result<Vec<DataSourceInfo>>

Stores all user data source server names and descriptions in a Vec

impl<V: Version> Environment<V>

pub fn new() -> Result<Environment<V>, Option<DiagnosticRecord>>

Creates an ODBC Environment and declares specification of V are used. You can use the shorthand create_environment_v3() instead.

Example

use odbc::*;
fn do_database_stuff() -> std::result::Result<(), Option<DiagnosticRecord>> {
    let env : Environment<Version3> = Environment::new()?; // first thing to do
    // ...
    Ok(())
}

Return

While most functions in this crate return a DiagnosticRecord in the event of an Error the creation of an environment is special. Since DiagnosticRecords are created using the environment, at least its allocation has to be successful to obtain one. If the allocation fails it is sadly not possible to receive further Diagnostics. Setting an unsupported version may however result in an ordinary Some(DiagnosticRecord).

impl Environment<Version3>

pub fn connect<'env>(
&'env self,
dsn: &str,
usr: &str,
pwd: &str
) -> Result<Connection<'env, AutocommitOn>>

Connects to an ODBC data source

Arguments

dsn - Data source name configured in the odbc.ini file

usr - User identifier

pwd - Authentication (usually password)

pub fn connect_with_connection_string<'env>(
&'env self,
connection_str: &str
) -> Result<Connection<'env, AutocommitOn>>

Connects to an ODBC data source using a connection string

See [SQLDriverConnect][1] for the syntax. [1]: https://docs.microsoft.com/en-us/sql/odbc/reference/syntax/sqldriverconnect-function