Struct odbc::Statement

pub struct Statement<'a, 'b, S, R, AC: AutocommitMode> { /* fields omitted */ }

A Statement can be used to execute queries and retrieves results.

Methods

impl<'a, 'b, S, R, AC: AutocommitMode> Statement<'a, 'b, S, R, AC>

pub fn bind_parameter<'c, T>(
self,
parameter_index: u16,
value: &'c T
) -> Result<Statement<'a, 'c, S, R, AC>> where
T: OdbcType<'c>,
T: ?Sized,
'b: 'c,

Binds a parameter to a parameter marker in an SQL statement.

Result

This method will destroy the statement and create a new one which may not outlive the bound parameter. This is to ensure that the statement will not dereference an invalid pointer during execution.

Arguments

parameter_index - Index of the marker to bind to the parameter. Starting at 1

value - Reference to bind to the marker

Example

let env = create_environment_v3().map_err(|e| e.unwrap())?;
let conn = env.connect("TestDataSource", "", "")?;
let stmt = Statement::with_parent(&conn)?;
let param = 1968;
let stmt = stmt.bind_parameter(1, &param)?;
let sql_text = "SELECT TITLE FROM MOVIES WHERE YEAR = ?";
if let Data(mut stmt) = stmt.exec_direct(sql_text)? {
    // ...
}

pub fn reset_parameters(self) -> Result<Statement<'a, 'a, S, R, AC>>

Releasing all parameter buffers set by bind_parameter. This method consumes the statement and returns a new one those lifetime is no longer limited by the buffers bound.

impl<'a, 'b, AC: AutocommitMode> Statement<'a, 'b, Allocated, NoResult, AC>

pub fn prepare(
self,
sql_text: &str
) -> Result<Statement<'a, 'b, Prepared, NoResult, AC>>

Prepares a statement for execution. Executing a prepared statement is faster than directly executing an unprepared statement, since it is already compiled into an Access Plan. This makes preparing statement a good idea if you want to repeatedly execute a query with a different set of parameters and care about performance.

Example

let env = create_environment_v3().map_err(|e| e.unwrap())?;
let conn = env.connect("TestDataSource", "", "")?;
let stmt = Statement::with_parent(&conn)?;
let mut stmt = stmt.prepare("SELECT TITLE FROM MOVIES WHERE YEAR = ?")?;

fn print_one_movie_from<'a> (year: u16, stmt: Statement<'a,'a, Prepared, NoResult, safe::AutocommitOn>) -> Result<Statement<'a, 'a, Prepared, NoResult, safe::AutocommitOn>>{
   let stmt = stmt.bind_parameter(1, &year)?;
   let stmt = if let Data(mut stmt) = stmt.execute()?{
       if let Some(mut cursor) = stmt.fetch()?{
           println!("{}", cursor.get_data::<String>(1)?.unwrap());
       }
       stmt.close_cursor()?
   } else {
      panic!("SELECT statement returned no result set");
   };
   stmt.reset_parameters()
};

for year in 1990..2010{
    stmt = print_one_movie_from(year, stmt)?
}

pub fn prepare_bytes(
self,
bytes: &[u8]
) -> Result<Statement<'a, 'b, Prepared, NoResult, AC>>

Prepares a statement for execution. Executing a prepared statement is faster than directly executing an unprepared statement, since it is already compiled into an Access Plan. This makes preparing statement a good idea if you want to repeatedly execute a query with a different set of parameters and care about performance.

Example

let env = create_environment_v3().map_err(|e| e.unwrap())?;
let conn = env.connect("TestDataSource", "", "")?;
let stmt = Statement::with_parent(&conn)?;
// need encode_rs crate
// let mut stmt = stmt.prepare_bytes(&GB2312.encode("select '你好' as hello").0)?;

impl<'a, 'b, AC: AutocommitMode> Statement<'a, 'b, Prepared, NoResult, AC>

pub fn num_result_cols(&self) -> Result<i16>

The number of columns in a result set

Can be called successfully only when the statement is in the prepared, executed, or positioned state. If the statement does not return columns the result will be 0.

pub fn describe_col(&self, idx: u16) -> Result<ColumnDescriptor>

Returns description struct for result set column with a given index. Note: indexing is starting from 1.

pub fn execute(self) -> Result<ResultSetState<'a, 'b, Prepared, AC>>

Executes a prepared statement.

impl<'a, 'b, 'env, AC: AutocommitMode> Statement<'a, 'b, Allocated, NoResult, AC>

pub fn with_parent(ds: &'a Connection<'env, AC>) -> Result<Self>

pub fn affected_row_count(&self) -> Result<SQLLEN>

pub fn tables(
self,
catalog_name: &String,
schema_name: &String,
table_name: &String,
table_type: &String
) -> Result<Statement<'a, 'b, Executed, HasResult, AC>>

pub fn tables_str(
self,
catalog_name: &str,
schema_name: &str,
table_name: &str,
table_type: &str
) -> Result<Statement<'a, 'b, Executed, HasResult, AC>>

pub fn tables_opt_str(
self,
catalog_name: Option<&str>,
schema_name: Option<&str>,
table_name: Option<&str>,
table_type: &str
) -> Result<Statement<'a, 'b, Executed, HasResult, AC>>

pub fn exec_direct(
self,
statement_text: &str
) -> Result<ResultSetState<'a, 'b, Executed, AC>>

Executes a preparable statement, using the current values of the parameter marker variables if any parameters exist in the statement.

SQLExecDirect is the fastest way to submit an SQL statement for one-time execution.

pub fn exec_direct_bytes(
self,
bytes: &[u8]
) -> Result<ResultSetState<'a, 'b, Executed, AC>>

Executes a preparable statement, using the current values of the parameter marker variables if any parameters exist in the statement.

SQLExecDirect is the fastest way to submit an SQL statement for one-time execution.

impl<'a, 'b, S, AC: AutocommitMode> Statement<'a, 'b, S, HasResult, AC>

pub fn affected_row_count(&self) -> Result<SQLLEN>

pub fn num_result_cols(&self) -> Result<i16>

The number of columns in a result set

Can be called successfully only when the statement is in the prepared, executed, or positioned state. If the statement does not return columns the result will be 0.

pub fn describe_col(&self, idx: u16) -> Result<ColumnDescriptor>

Returns description struct for result set column with a given index. Note: indexing is starting from 1.

pub fn fetch<'s>(&'s mut self) -> Result<Option<Cursor<'s, 'a, 'b, S, AC>>>

Fetches the next rowset of data from the result set and returns data for all bound columns.

pub fn close_cursor(self) -> Result<Statement<'a, 'b, S, NoResult, AC>>

Call this method to reuse the statement to execute another query.

For many drivers allocating new statements is expensive. So reusing a Statement is usually more efficient than freeing an existing and allocating a new one. However to reuse a statement any open result sets must be closed. Only call this method if you have already read the result set returned by the previous query, or if you do no not intend to read it.

Example

let env = create_environment_v3().map_err(|e| e.unwrap())?;
let conn = env.connect("TestDataSource", "", "")?;
let stmt = Statement::with_parent(&conn)?;
let stmt = match stmt.exec_direct("CREATE TABLE STAGE (A TEXT, B TEXT);")?{
    // Some drivers will return an empty result set. We need to close it before we can use
    // statement again.
    Data(stmt) => stmt.close_cursor()?,
    NoData(stmt) => stmt,
};
let stmt = stmt.exec_direct("INSERT INTO STAGE (A, B) VALUES ('Hello', 'World');")?;
//...