Struct odbc::Statement

pub struct Statement<'a, 'b, S, R> { /* fields omitted */ }

A Statement can be used to execute queries and retrieves results.

Methods

impl<'a, 'b, S, R> Statement<'a, 'b, S, R>

fn bind_parameter<'c, T: ?Sized>(self,
parameter_index: u16,
value: &'c T)
-> Result<Statement<'a, 'c, S, R>> where T: InputParameter, T: , 'b: 'c

Binds a parameter to a parameter marker in an SQL statement.

Result

This method will destroy the statement and create a new one which may not outlive the bound parameter. This is to ensure that the statement will not derefernce an invalid pointer during execution.

Arguments

parameter_index - Index of the marker to bind to the parameter. Starting at 1

value - Reference to bind to the marker

Example

let env = Environment::new()?.set_odbc_version_3()?;
let conn = DataSource::with_parent(&env)?.connect("TestDataSource", "", "")?;
let stmt = Statement::with_parent(&conn)?;
let param = 1968;
let stmt = stmt.bind_parameter(1, &param)?;
let sql_text = "SELECT TITLE FROM MOVIES WHERE YEAR = ?";
if let Data(mut stmt) = stmt.exec_direct(sql_text)? {
    // ...
}

fn reset_parameters(self) -> Result<Statement<'a, 'a, S, R>>

Releasing all parameter buffers set by bind_parameter. This method consumes the statement and returns a new one those lifetime is no longer limited by the buffers bound.

impl<'a, 'b> Statement<'a, 'b, Allocated, NoResult>

fn prepare(self,
sql_text: &str)
-> Result<Statement<'a, 'b, Prepared, NoResult>>

Prepares a statement for execution. Executing a prepared statement is faster than directly executing an unprepared statement, since it is already compiled into an Access Plan. This makes preparing statement a good idea if you want to repeatedly execute a query with a different set of parameters and care about performance.

Example

let env = Environment::new().unwrap().set_odbc_version_3()?;
let conn = DataSource::with_parent(&env)?.connect("TestDataSource", "", "")?;
let stmt = Statement::with_parent(&conn)?;
let mut stmt = stmt.prepare("SELECT TITLE FROM MOVIES WHERE YEAR = ?")?;

fn print_one_movie_from<'a> (year: u16, stmt: Statement<'a,'a, Prepared, NoResult>) -> Result<Statement<'a, 'a, Prepared, NoResult>>{
   let stmt = stmt.bind_parameter(1, &year)?;
   let stmt = if let Data(mut stmt) = stmt.execute()?{
       if let Some(mut cursor) = stmt.fetch()?{
           println!("{}", cursor.get_data::<String>(1)?.unwrap());
       }
       stmt.close_cursor()?
   } else {
      panic!("SELECT statement returned no result set");
   };
   stmt.reset_parameters()
};

for year in 1990..2010{
    stmt = print_one_movie_from(year, stmt)?
}

impl<'a, 'b> Statement<'a, 'b, Prepared, NoResult>

fn execute(self) -> Result<ResultSetState<'a, 'b, Prepared>>

Executes a prepared statement.

impl<'a, 'b> Statement<'a, 'b, Allocated, NoResult>

fn with_parent(ds: &'a DataSource<Connected>) -> Result<Self>

fn tables(self) -> Result<Statement<'a, 'b, Executed, HasResult>>

fn exec_direct(self,
statement_text: &str)
-> Result<ResultSetState<'a, 'b, Executed>>

Executes a preparable statement, using the current values of the parameter marker variables if any parameters exist in the statement.

SQLExecDirect is the fastest way to submit an SQL statement for one-time execution.

impl<'a, 'b, S> Statement<'a, 'b, S, HasResult>

fn num_result_cols(&self) -> Result<i16>

The number of columns in a result set

Can be called successfully only when the statement is in the prepared, executed, or positioned state. If the statement does not return columns the result will be 0.

fn fetch<'c>(&'c mut self) -> Result<Option<Cursor<'c, 'a, 'b, S>>>

Fetches the next rowset of data from the result set and returns data for all bound columns.

fn close_cursor(self) -> Result<Statement<'a, 'b, S, NoResult>>

Call this method to reuse the statement to execute another query.

For many drivers allocating new statemens is expensive. So reusing a Statement is usually more efficient than freeing an existing and alloctaing a new one. However to reuse a statement any open result sets must be closed. Only call this method if you have already read the result set returned by the previous query, or if you do no not intend to read it.

Example

let env = Environment::new().unwrap().set_odbc_version_3()?;
let conn = DataSource::with_parent(&env)?.connect("TestDataSource", "", "")?;
let stmt = Statement::with_parent(&conn)?;
let stmt = match stmt.exec_direct("CREATE TABLE STAGE (A TEXT, B TEXT);")?{
    // Some drivers will return an empty result set. We need to close it before we can use
    // statement again.
    Data(stmt) => stmt.close_cursor()?,
    NoData(stmt) => stmt,
};
let stmt = stmt.exec_direct("INSERT INTO STAGE (A, B) VALUES ('Hello', 'World');")?;
//...